+++ src/components/Sidebar/SidebarItemText/SidebarItemText.styled.tsx
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import { SidebarItemTextProps } from "./SidebarItemText";
 
 export const SidebarItemTextStyled = styled.p`
     margin             : 0;